requirements.txt
streamlit==1.42.0
pandas==2.2.2
gspread==6.1.2
google-auth==2.33.0
google-api-python-client==2.142.0
